chore(Modified): Delegate Extensions
