+++ Unit5GroupProjectOne/View Controllers/SearchViewController.swift
@@ -5,22 +5,21 @@
 //  Created by C4Q on 1/18/18.
 //  Copyright © 2018 C4Q. All rights reserved.
 //
-
 import UIKit
 import SnapKit
 import MapKit
 
 class SearchViewController: UIViewController {
     
-    private let cellSpacing: CGFloat = UIScreen.main.bounds.width * 0.02
+     let cellSpacing: CGFloat = UIScreen.main.bounds.width * 0.02
     
-    private let searchView = SearchView()
+     let searchView = SearchView()
     
-    private lazy var venueSearchBarController = UISearchController(searchResultsController: nil)
+     lazy var venueSearchBarController = UISearchController(searchResultsController: nil)
     
-    private var venues: [Venue] = [] {
+     var venues: [Venue] = [] {
         didSet {
-            searchView.venueCollectionView.reloadData()
+            searchView.venueCollectionView.reloadData()            
             
             //should create a list of annotations from these venues
             //should assign that list to self.annotations (replacing the old list!)
@@ -29,23 +28,17 @@ class SearchViewController: UIViewController {
             
             for venue in venues {
                 let annotation = MKPointAnnotation()
-                
                 annotation.coordinate = CLLocationCoordinate2D(latitude: venue.location.lat, longitude: venue.location.lng)
-                
                 annotation.title = venue.name
                 annotation.subtitle = venue.categories.map{$0.shortName}.joined(separator: ", ")
-                
                 annotations.append(annotation)
             }
-            
             searchView.mapView.removeAnnotations(self.annotations)
-            
             self.annotations = annotations
-            
         }
     }
     
-    private var annotations: [MKAnnotation] = [] {
+     var annotations: [MKAnnotation] = [] {
         didSet {
             //when this changes
                 //should remove the old pins from the map
@@ -76,7 +69,7 @@ class SearchViewController: UIViewController {
 //MARK: - Helper Functions
 extension SearchViewController {
     
-    private func configureNavBar() {
+     func configureNavBar() {
         //TODO: Edit title
         navigationItem.title = "Venue Search"
         navigationController?.navigationBar.prefersLargeTitles = true
@@ -96,7 +89,7 @@ extension SearchViewController {
         self.present(navController, animated: true, completion: nil)
     }
     
-    private func delegateAndDataSource() {
+     func delegateAndDataSource() {
         navigationItem.searchController?.searchBar.delegate = self
         searchView.locationSearchBar.delegate = self
         searchView.venueCollectionView.delegate = self
@@ -107,7 +100,7 @@ extension SearchViewController {
         
     }
     
-    private func constrainView() {
+     private func constrainView() {
         view.addSubview(searchView)
         searchView.backgroundColor = .cyan
         searchView.snp.makeConstraints { (view) in
@@ -115,7 +108,7 @@ extension SearchViewController {
         }
     }
     
-    private func setUpLocationServices() {
+     func setUpLocationServices() {
         let locationService = LocationService.manager.checkAuthorizationStatusAndLocationServices()
         
         if locationService.locationServicesEnabled {
@@ -125,22 +118,16 @@ extension SearchViewController {
             case .restricted:
                 let alertController = UIAlertController(title: "Warning", message: "This app is not authorized to use location services.", preferredStyle: .alert)
                 let alertAction = UIAlertAction(title: "OK", style: .default, handler: nil)
-                
                 alertController.addAction(alertAction)
-                
                 self.present(alertController, animated: true, completion: nil)
-                
             default:
                 //stuff that you might want to happen if the app starts and the user access is on
                 break
-
             }
         }
     }
-    
-    private func presentSettingsAlertController(withTitle title: String?, andMessage message: String?) {
+     func presentSettingsAlertController(withTitle title: String?, andMessage message: String?) {
         let alertController = UIAlertController(title: title, message: message, preferredStyle: .alert)
-        
         let cancelAction = UIAlertAction(title: "Cancel", style: .cancel, handler: nil)
         let settingsAction = UIAlertAction(title: "Settings", style: .default, handler: { (_) in
             
@@ -149,28 +136,22 @@ extension SearchViewController {
             }
             
             UIApplication.shared.open(settingsURL, options: [:], completionHandler: nil)
-            
         })
-        
         alertController.addAction(cancelAction)
         alertController.addAction(settingsAction)
-        
         self.present(alertController, animated: true, completion: nil)
     }
     
-    private func createAlertController(withTitle title: String?, andMessage message: String?) -> UIAlertController {
+     func createAlertController(withTitle title: String?, andMessage message: String?) -> UIAlertController {
         
         let alertController = UIAlertController(title: title, message: message, preferredStyle: .alert)
         let alertAction = UIAlertAction(title: "OK", style: .default, handler: nil)
         alertController.addAction(alertAction)
-        
         return alertController
     }
     
-    private func getVenues(fromSearchTerm searchTerm: String, latitude: Double, andLongitude longitude: Double) {
-        
+     func getVenues(fromSearchTerm searchTerm: String, latitude: Double, andLongitude longitude: Double) {
         VenueAPIClient.manager.getVenues(lat: latitude, lon: longitude, search: searchTerm, completion: { (venues) in
-            
             if venues.isEmpty {
                 print("It's empty!!!")
                 
@@ -181,117 +162,17 @@ extension SearchViewController {
                 
                 return
             }
-            
             print(venues)
             self.venues = venues
-            
         }, errorHandler: { (error) in
             
             //TODO: Present the alert
             let alertController = self.createAlertController(withTitle: "Error", andMessage: "An error occurred:\n\(error)")
-            
             self.present(alertController, animated: true, completion: nil)
             
         })
-
     }
-    
 }
-
-//MARK: - Location Service Delegate Methods
-extension SearchViewController: LocationServiceDelegate {
-    
-    //Melissa to QA: Please check if this doesn't make the alert pop up too often! Thanks!!
-    func locationServiceAuthorizationStatusChanged(toStatus status: CLAuthorizationStatus) {
-        print(status)
-        
-        switch status {
-        case .denied:
-            if self.presentedViewController == nil {
-                presentSettingsAlertController(withTitle: "Location Services Not Enabled", andMessage: "Please enable Location Services in Settings for better search results.")
-            }
-        case .restricted:
-            let alertController = UIAlertController(title: "Warning", message: "This app is not authorized to use location services.", preferredStyle: .alert)
-            let alertAction = UIAlertAction(title: "OK", style: .default, handler: nil)
-            
-            alertController.addAction(alertAction)
-            
-            self.present(alertController, animated: true, completion: nil)
-            
-        default:
-            break
-        }
-        
-    }
-    
-    func userLocationUpdateFailed(withError error: Error) {
-        print(error)
-        presentSettingsAlertController(withTitle: "Could Not Get User Location", andMessage: "Please enable Location Services in Settings or check network connectivity.")
-    }
-    
-    func userLocationUpdatedToLocation(_ location: CLLocation) {
-        let userCoordinates = searchView.mapView.userLocation.coordinate
-        let mapSpan = MKCoordinateSpan(latitudeDelta: 0.1, longitudeDelta: 0.1)
-        searchView.mapView.showsUserLocation = true
-        let userRegion = MKCoordinateRegion(center: CLLocationCoordinate2D(latitude: userCoordinates.latitude, longitude: userCoordinates.longitude), span: mapSpan)
-        searchView.mapView.setRegion(userRegion, animated: true)
-        //TODO: set the place holder everytime the current location changes!! - use the geocoder to find the actual place!
-        //TODO: fix user tracking button
-    }
-    
-}
-
-//MARK: - Map View Delegate Methods
-//TODO: Make Map View Delegate Methods
-//should have a map manager or something? need a map delegate!!
-//should check the map view's view that is return for each annotation!
-
-extension SearchViewController: MKMapViewDelegate {
-    //to do!!
-    
-    func mapView(_ mapView: MKMapView, didFailToLocateUserWithError error: Error) {
-        print(error)
-        //maybe present error message about location not enabled???
-    }
-    
-    func mapView(_ mapView: MKMapView, annotationView view: MKAnnotationView, calloutAccessoryControlTapped control: UIControl) {
-        print("tapped accessory!!!")
-        //should segue to detail view
-    }
-    
-    func mapView(_ mapView: MKMapView, viewFor annotation: MKAnnotation) -> MKAnnotationView? {
-        //returns the view for each annotation - the bubble thing that pops up
-        
-        //if annotation is the user location, don't do anything
-        if annotation is MKUserLocation {
-            return nil
-        }
-        
-        let annotationView = mapView.dequeueReusableAnnotationView(withIdentifier: "mapAnnotationView") as? MKMarkerAnnotationView
-        
-        //the bubble that pops up when you click the annotation
-        annotationView?.canShowCallout = true
-        annotationView?.annotation = annotation
-        
-        annotationView?.animatesWhenAdded = true
-        
-        annotationView?.rightCalloutAccessoryView = UIButton(type: .detailDisclosure)
-        
-        //TODO: MELISSA CHANGE TO LIGHTER PUPRLE
-        annotationView?.markerTintColor = UIColor.purple
-
-        //TODO: NATE DO SHADOWS!!
-//        annotationView?.layer.shadowColor
-//        annotationView?.layer.shadowOffset
-//        annotationView?.layer.shadowRadius
-        
-        annotationView?.displayPriority = .required
-        
-        return annotationView
-    }
-    
-}
-
 
 //MARK: - Collection View Delegate Flow Layout Methods
 extension SearchViewController: UICollectionViewDelegateFlowLayout {
@@ -316,27 +197,18 @@ extension SearchViewController: UICollectionViewDelegateFlowLayout {
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, insetForSectionAt section: Int) -> UIEdgeInsets {
         return UIEdgeInsets(top: 0, left: cellSpacing, bottom: 0, right: cellSpacing)
     }
-    
 }
 
 //MARK: - Collection View Delegate and Data Source Methods
 extension SearchViewController: UICollectionViewDelegate, UICollectionViewDataSource {
     
     func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        
         let selectedAnnotation = annotations[indexPath.row]
-        
         let center = CLLocationCoordinate2D(latitude: selectedAnnotation.coordinate.latitude, longitude: selectedAnnotation.coordinate.longitude)
-        
         let span = MKCoordinateSpan(latitudeDelta: 0.05, longitudeDelta: 0.05)
-        
         let region = MKCoordinateRegion(center: center, span: span)
-        
         searchView.mapView.setRegion(region, animated: true)
-        
         searchView.mapView.selectAnnotation(selectedAnnotation, animated: true)
-        
-        
     }
     
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
@@ -346,77 +218,8 @@ extension SearchViewController: UICollectionViewDelegate, UICollectionViewDataSo
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
         let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "SearchCollectionCell", for: indexPath) as! SearchCollectionViewCell
         cell.contentView.backgroundColor = .gray
-        
         return cell
     }
-    
 }
 
-//TODO: Finish Search Bar Delegate
-    //tapping search should get the list of venues, and changing the list of venues
-extension SearchViewController: UISearchBarDelegate {
-    
-    func searchBarSearchButtonClicked(_ searchBar: UISearchBar) {
-        
-        //probably should set up user preferences to save last search and last location (lol)
-        
-        guard
-            let venueSearchText = venueSearchBarController.searchBar.text,
-            !venueSearchText.isEmpty,
-            let formattedVenueSearchText = venueSearchText.addingPercentEncoding(withAllowedCharacters: .urlHostAllowed)
-            else {
 
-            return
-        }
-        
-        guard
-            let locationSearchText = searchView.locationSearchBar.text,
-            !locationSearchText.isEmpty
-            else {
-            
-                let locationService = LocationService.manager.checkAuthorizationStatusAndLocationServices()
-                
-                //if location services is enabled - do search
-                if locationService.locationServicesEnabled && (locationService.authorizationStatus == .authorizedAlways || locationService.authorizationStatus == .authorizedWhenInUse) {
-                    //if user location access is allowed
-                    
-                    //do search
-                    
-                    let userLocation = searchView.mapView.userLocation
-                    
-                    getVenues(fromSearchTerm: formattedVenueSearchText, latitude: userLocation.coordinate.latitude, andLongitude: userLocation.coordinate.longitude)
-                    
-                } else {
-                    //if location services is not enabled - present alert controller
-                    presentSettingsAlertController(withTitle: "Location Services Not Enabled", andMessage: "Please enable Location Services in Settings for better search results.")
-                }
-            
-            return
-        }
-        
-        //if user enters location
-        LocationService.manager.getLatAndLong(fromLocation: locationSearchText) { (error, coordinate) in
-        
-            if let error = error {
-                
-                //present alert that says could not get coordinates from user inputted location
-                let alertController = self.createAlertController(withTitle: "Error", andMessage: "Could not get coordinates from user-input location.\n\"\(error.localizedDescription)\"")
-                
-                self.present(alertController, animated: true, completion: nil)
-                
-                return
-            }
-            
-            if let coordinate = coordinate {
-                
-                self.getVenues(fromSearchTerm: formattedVenueSearchText, latitude: coordinate.latitude, andLongitude: coordinate.longitude)
-                
-            }
-            
-        }
-        
-        searchBar.resignFirstResponder()
-        
-    }
-    
-}
